Model Driven Form
The model-driven approach is another way to handle building forms in
Angular. It has a lot of similarities to the template-driven approach,
but the big difference is the form is built in the component class.
That allows you to have a contract that your form needs to adhere to,
the ability to set up validation rules, subscribe to field value
changes, and ultimately unit test your form logic, without any UI
layer. So template-driven gives you an easy way to get up and running,
whereas model-driven allows you to take your forms wherever they need
to go.